add parent and children to MTA
Parent and children structures have been added to MTA so that each node has a copy of who their parent and children are if applicable. In addition, MTA reconvergence restarts the send process from the downstream broken node, which needs additional work, but does work as written for the time being.